app.py
!pip install torch torchvision transformers nltk

import torch
import torchvision.transforms as transforms
from torch.utils.data import DataLoader
from transformers import AutoTokenizer, AutoModelForSeq2SeqLM
import nltk
from nltk.tokenize import word_tokenize
import os
import zipfile

# Download dataset
!wget http://images.cocodataset.org/zips/train2017.zip

with zipfile.ZipFile("train2017.zip", 'r') as zip_ref:
    zip_ref.extractall("train2017")

!wget http://images.cocodataset.org/annotations/annotations_trainval2017.zip
!unzip annotations_trainval2017.zip -d annotations

# Preprocessing
# Simulating preprocessing for demonstration
transform = transforms.Compose([
    transforms.Resize((256, 256)),
    transforms.ToTensor(),
])

def preprocess_image(image_path):
    from PIL import Image
    image = Image.open(image_path)
    return transform(image).unsqueeze(0)

# Loading a pre-trained model
def load_model():
    tokenizer = AutoTokenizer.from_pretrained("facebook/bart-large-cnn")
    model = AutoModelForSeq2SeqLM.from_pretrained("facebook/bart-large-cnn")
    return tokenizer, model

# Generating captions
def generate_caption(image_tensor, tokenizer, model):
    # Placeholder logic for caption generation
    input_ids = tokenizer.encode("Example caption generation", return_tensors="pt")
    summary_ids = model.generate(input_ids, max_length=50, min_length=5, length_penalty=2.0, num_beams=4, early_stopping=True)
    return tokenizer.decode(summary_ids[0], skip_special_tokens=True)

# Main execution
if __name__ == "__main__":
    tokenizer, model = load_model()
    test_image = "train2017/sample.jpg"  # Replace with a valid image path
    preprocessed_image = preprocess_image(test_image)
    caption = generate_caption(preprocessed_image, tokenizer, model)
    print("Generated Caption:", caption)
