requirements.txt
streamlit
torch
torchvision
transformers
nltk
Pillow
